Fix buffer overflow (#27)
* Fix buffer overflow.

Before this, the amqp messages would get copied to
ring buffer without checking, if there is enough space
in the buffer. That would lead to buffer overflows for
long messages. Now the sizes are checked and if the
message is longer than the buffer, it's first read
from the bus and then it's discarded.

* Committing clang-format changes

---------

Co-authored-by: InfraWatch CI <robot@infra.watch>